+++ src/transport/emitter.ts
@@ -1,6 +1,6 @@
 import { CloudEvent } from "../event/cloudevent";
 import { emitBinary, emitStructured } from "./http";
-import { Protocol } from ".";
+import { Protocol } from "./protocols";
 import { AxiosResponse } from "axios";
 import { Agent } from "http";
