--- src/transport/index.ts
@@ -1,4 +1,0 @@
-export * from "./emitter";
-export * from "./receiver";
-export * as http from "./http/headers";
-export * from "./protocols";
